requirements.txt
Flask==3.1.3
Werkzeug==3.1.8
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.3
blinker==1.9.0
python-dotenv==1.0.1
bcrypt==4.2.1

requests==2.32.3